Refactor monitor loopback capture service
