+++ src/qubex/measurement/services/__init__.py
@@ -3,6 +3,7 @@
 from .measurement_amplification_service import MeasurementAmplificationService
 from .measurement_classification_service import MeasurementClassificationService
 from .measurement_execution_service import MeasurementExecutionService
+from .measurement_monitor_service import MeasurementMonitorService
 from .measurement_session_service import MeasurementSessionService
 from .measurement_stability_service import MeasurementStabilityService
 
@@ -10,6 +11,7 @@ __all__ = [
     "MeasurementAmplificationService",
     "MeasurementClassificationService",
     "MeasurementExecutionService",
+    "MeasurementMonitorService",
     "MeasurementSessionService",
     "MeasurementStabilityService",
 ]
